Use relative imports and import submodules in __init__.py
