+++ sampledbapi/actiontypes.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from sampledbapi import SampleDBObject, get_data
+from .comm import SampleDBObject, get_data
 
 __all__ = ["ActionType", "get_list", "get"]
